[Prover Service] Remove dead code from types directory.
