+++ prover-service/Cargo.toml
@@ -26,7 +26,6 @@ aptos-types = { workspace = true }
 ark-bn254 = { workspace = true }
 ark-ff = { workspace = true }
 ark-groth16 = { workspace = true }
-ark-serialize = { workspace = true }
 async-trait = { workspace = true }
 base64 = { workspace = true }
 bcs = { workspace = true }
@@ -37,7 +36,6 @@ http = { workspace = true }
 hyper = { workspace = true }
 jsonwebtoken = { workspace = true }
 num-bigint = { workspace = true }
-num-traits = { workspace = true }
 once_cell = { workspace = true }
 prometheus = { workspace = true }
 rand = { workspace = true }
