+++ prover-service/src/request_handler/mod.rs
@@ -4,3 +4,4 @@ pub mod deployment_information;
 pub mod handler;
 pub mod prover_handler;
 pub mod prover_state;
+pub mod types;
