+++ prover-service/src/request_handler/prover_handler.rs
@@ -9,27 +9,19 @@ use crate::metrics::{
     PROOF_VERIFICATION_LABEL, PROVER_RESPONSE_GENERATION_LABEL, VALIDATE_PROVE_REQUEST_LABEL,
     WITNESS_GENERATION_LABEL,
 };
-use crate::request_handler::handler;
+use crate::request_handler::types::{ProverServiceResponse, RequestInput};
+use crate::request_handler::{handler, types};
 use crate::{
-    input_processing, metrics,
-    request_handler::prover_state::ProverServiceState,
-    training_wheels,
-    types::api::{ProverServiceResponse, RequestInput},
+    input_processing, metrics, request_handler::prover_state::ProverServiceState, training_wheels,
     utils,
 };
 use aptos_keyless_common::input_processing::circuit_input_signals::{CircuitInputSignals, Padded};
 use aptos_keyless_common::PoseidonHash;
 use aptos_logger::{error, warn};
-use aptos_types::keyless::{g1_projective_str_to_affine, g2_projective_str_to_affine};
-use aptos_types::{
-    keyless::{G1Bytes, G2Bytes, Groth16Proof},
-    transaction::authenticator::EphemeralSignature,
-};
-use ark_bn254::Bn254;
+use aptos_types::keyless::Groth16Proof;
+use aptos_types::transaction::authenticator::EphemeralSignature;
 use ark_ff::PrimeField;
-use ark_groth16::{PreparedVerifyingKey, VerifyingKey};
 use hyper::{Body, Request, Response, StatusCode};
-use serde::{Deserialize, Serialize};
 use std::convert::Infallible;
 use std::fs;
 use std::process::Command;
@@ -272,7 +264,7 @@ async fn generate_groth16_proof(
     };
 
     // Encode the rapidsnark proof into a groth16 proof
-    let groth16_proof = match encode_proof(&rapidsnark_proof_response) {
+    let groth16_proof = match types::encode_proof(&rapidsnark_proof_response) {
         Ok(proof) => proof,
         Err(error) => {
             return Err(ProverServiceError::UnexpectedError(format!(
@@ -286,7 +278,7 @@ async fn generate_groth16_proof(
     let verification_key_file_path = prover_service_state
         .prover_service_config()
         .verification_key_file_path();
-    let groth16_prepare_verifying_key = prepared_vk(&verification_key_file_path)?;
+    let groth16_prepare_verifying_key = types::prepared_vk(&verification_key_file_path)?;
 
     // Update the proof deserialization metrics
     metrics::update_prove_request_breakdown_metrics(
@@ -451,113 +443,6 @@ async fn validate_prove_request_input(
     );
 
     Ok(verified_input)
-}
-
-// TODO: should we rename RawVK?
-
-/// A raw VK as outputted by circom in YAML format
-#[allow(non_snake_case)]
-#[derive(Serialize, Deserialize, Debug)]
-struct RawVK {
-    vk_alpha_1: Vec<String>,
-    vk_beta_2: Vec<Vec<String>>,
-    vk_gamma_2: Vec<Vec<String>>,
-    vk_delta_2: Vec<Vec<String>>,
-    IC: Vec<Vec<String>>,
-}
-
-/// This function uses the decimal uncompressed point serialization which is outputted by circom
-pub fn prepared_vk(vk_file_path: &str) -> Result<PreparedVerifyingKey<Bn254>, ProverServiceError> {
-    // Fetch the raw VK from the file
-    let raw_vk_yaml = utils::read_string_from_file_path(vk_file_path);
-    let raw_vk: RawVK = match serde_yaml::from_str(&raw_vk_yaml) {
-        Ok(raw_vk) => raw_vk,
-        Err(error) => {
-            return Err(ProverServiceError::UnexpectedError(format!(
-                "Failed to parse VK YAML file: {}! Error: {}",
-                vk_file_path, error
-            )));
-        }
-    };
-
-    // Construct alpha_g1
-    let alpha_g1 =
-        g1_projective_str_to_affine(&raw_vk.vk_alpha_1[0], &raw_vk.vk_alpha_1[1]).unwrap();
-
-    // Construct beta_g2
-    let beta_g2 = g2_projective_str_to_affine(
-        [&raw_vk.vk_beta_2[0][0], &raw_vk.vk_beta_2[0][1]],
-        [&raw_vk.vk_beta_2[1][0], &raw_vk.vk_beta_2[1][1]],
-    )
-    .unwrap();
-
-    // Construct gamma_g2
-    let gamma_g2 = g2_projective_str_to_affine(
-        [&raw_vk.vk_gamma_2[0][0], &raw_vk.vk_gamma_2[0][1]],
-        [&raw_vk.vk_gamma_2[1][0], &raw_vk.vk_gamma_2[1][1]],
-    )
-    .unwrap();
-
-    // Construct delta_g2
-    let delta_g2 = g2_projective_str_to_affine(
-        [&raw_vk.vk_delta_2[0][0], &raw_vk.vk_delta_2[0][1]],
-        [&raw_vk.vk_delta_2[1][0], &raw_vk.vk_delta_2[1][1]],
-    )
-    .unwrap();
-
-    // Construct gamma_abc_g1
-    let mut gamma_abc_g1 = Vec::new();
-    for p in raw_vk.IC {
-        gamma_abc_g1.push(g1_projective_str_to_affine(&p[0], &p[1]).unwrap());
-    }
-
-    // Create and return the prepared verifying key
-    let vk = VerifyingKey {
-        alpha_g1,
-        beta_g2,
-        gamma_g2,
-        delta_g2,
-        gamma_abc_g1,
-    };
-    Ok(PreparedVerifyingKey::from(vk))
-}
-
-/// A rapidsnark proof response
-#[derive(Deserialize)]
-pub struct RapidsnarkProofResponse {
-    pi_a: [String; 3],
-    pi_b: [[String; 2]; 3],
-    pi_c: [String; 3],
-}
-
-impl RapidsnarkProofResponse {
-    fn pi_b_str(&self) -> [[&str; 2]; 3] {
-        [
-            [&self.pi_b[0][0], &self.pi_b[0][1]],
-            [&self.pi_b[1][0], &self.pi_b[1][1]],
-            [&self.pi_b[2][0], &self.pi_b[2][1]],
-        ]
-    }
-}
-
-/// Encodes a Rapidsnark proof response into a Groth16 proof
-pub fn encode_proof(
-    rapidsnark_proof_response: &RapidsnarkProofResponse,
-) -> Result<Groth16Proof, ProverServiceError> {
-    let new_pi_a = G1Bytes::new_unchecked(
-        &rapidsnark_proof_response.pi_a[0],
-        &rapidsnark_proof_response.pi_a[1],
-    )?;
-    let new_pi_b = G2Bytes::new_unchecked(
-        rapidsnark_proof_response.pi_b_str()[0],
-        rapidsnark_proof_response.pi_b_str()[1],
-    )?;
-    let new_pi_c = G1Bytes::new_unchecked(
-        &rapidsnark_proof_response.pi_c[0],
-        &rapidsnark_proof_response.pi_c[1],
-    )?;
-
-    Ok(Groth16Proof::new(new_pi_a, new_pi_b, new_pi_c))
 }
 
 /// Generates the witness file using the given prover config and circuit input signals
