+++ prover-service/src/tests/common/types.rs
@@ -6,11 +6,10 @@ use std::time::{SystemTime, UNIX_EPOCH};
 
 use super::{gen_test_ephemeral_pk, gen_test_ephemeral_pk_blinder, get_test_pepper};
 use crate::external_resources::prover_config::ProverServiceConfig;
+use crate::request_handler::types::{EphemeralPublicKeyBlinder, RequestInput};
 use crate::tests::common::get_config;
 use crate::{
-    input_processing::rsa::RsaPrivateKey,
-    training_wheels::verification_logic::compute_nonce,
-    types::api::{EphemeralPublicKeyBlinder, RequestInput},
+    input_processing::rsa::RsaPrivateKey, training_wheels::verification_logic::compute_nonce,
 };
 use aptos_keyless_common::input_processing::encoding::FromFr;
 use aptos_types::{
