+++ prover-service/src/training_wheels/mod.rs
@@ -8,8 +8,8 @@ use crate::external_resources::jwk_fetcher::{get_federated_jwk, JWKCache};
 use crate::external_resources::prover_config::ProverServiceConfig;
 use crate::input_processing::types::VerifiedInput;
 use crate::request_handler::prover_state::ProverServiceState;
+use crate::request_handler::types::RequestInput;
 use crate::training_wheels::verification_logic::compute_nonce;
-use crate::types::api::RequestInput;
 use anyhow::{anyhow, bail, ensure};
 use aptos_keyless_common::input_processing::encoding::{AsFr, DecodedJWT};
 use aptos_types::jwks::rsa::RSA_JWK;
