--- prover-service/src/types/mod.rs
@@ -1,4 +1,0 @@
-// Copyright (c) Aptos Foundation
-
-pub mod api;
-pub mod groth16_vk;
